+++ .gitignore
@@ -1,7 +1,18 @@
-**/node_modules
+**/web_visualizer/node_modules
+/web_visualizer/node_modules/**
 **/.vscode
+/.vscode/**
 **/venv
-**/.webassets-cache
+/venv/**
+**/web_visualizer/.webassets-cache
+/web_visualizer/.webassets-cache/**
 **/static/bundle.js
-**/data
-**/__pychache__
+/static/bundle.js/**
+**/web_visualizer/__pychache__
+/web_visualizer/__pychache__/**
+**/web_visualizer.egg-info
+/web_visualizer.egg-info/**
+**/web_visualizer/static/.webassets-cache
+/web_visualizer/static/.webassets-cache
+**/web_visualizer/data
+/web_visualizer/data/**
