+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='web_visualizer',
+    packages=['web_visualizer'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ]
+)
