+++ web_visualizer/static/js/animation.js
