+++ web_visualizer/static/js/helpers.js
