+++ web_visualizer/static/js/store.js
