+++ web_visualizer/static/styles/styles.css
